requirements.txt
# タスク管理ツール
python-frontmatter>=1.0.0

# データ処理
pandas>=2.0.0
numpy>=1.24.0

# 機械学習
scikit-learn>=1.3.0

# 深層学習・トランスフォーマー
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0

# その他
pyyaml>=6.0
tqdm>=4.65.0

